backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pymysql==1.1.1
python-dotenv==1.0.1
email-validator==2.1.0.post1
